commit remarks in cw 2
