+++ components/add-post-page-component.js
@@ -4,9 +4,8 @@ import { posts, goToPage } from "../index.js";
 import { renderUploadImageComponent } from "./upload-image-component.js";
 import { sanitize } from "../helpers.js";
 
-let imageUrl = "";
-
 export function renderAddPostPageComponent({ appEl, onAddPostClick }) {
+  let imageUrl = "";
   const render = () => {
     // TODO: Реализовать страницу добавления поста
     const appHtml = `
@@ -44,7 +43,12 @@ export function renderAddPostPageComponent({ appEl, onAddPostClick }) {
     renderUploadImageComponent({
       element: appEl.querySelector(".upload-image-container"),
       onImageUrlChange(newImageUrl) {
-        imageUrl = newImageUrl;
+        if (!newImageUrl.trim()) {
+          alert("Загрузите фото");
+          return;
+        } else {
+          imageUrl = newImageUrl.trim();
+        }
       },
     });
