+++ components/posts-page-component.js
@@ -1,6 +1,8 @@
 import { USER_POSTS_PAGE } from "../routes.js";
 import { renderHeaderComponent } from "./header-component.js";
-import { posts, goToPage, page, likePost, id, delPost } from "../index.js";
+import { posts, goToPage, page, id } from "../index.js";
+import { likePost, delPost } from "./toggle-del-posts-components.js"
+import { sanitize } from "../helpers.js"
 import { user } from "../index.js";
 import { formatDistance } from "date-fns";
 // Require Russian locale
@@ -18,7 +20,7 @@ export function renderPostsPageComponent({ appEl }) {
           page != USER_POSTS_PAGE
             ? `<div class="post-header" data-user-id=${post.user.id}>
             <img src=${post.user.imageUrl} class="post-header__user-image">
-            <p class="post-header__user-name">${post.user.name}</p>
+            <p class="post-header__user-name">${sanitize(post.user.name)}</p>
           </div>`
             : ""
         }
@@ -57,7 +59,7 @@ export function renderPostsPageComponent({ appEl }) {
             <div>
               <p class="post-text">
                 <span class="user-name">            
-                ${post.user.name}
+                ${sanitize(post.user.name)}
                 </span>
                 ${post.description}
               </p>
@@ -74,9 +76,7 @@ export function renderPostsPageComponent({ appEl }) {
                   ? `<button data-id=${post.id} class="button">
                 <div title="Удалить пост"></div>
                 Удалить</button>`
-                  : `<button data-id=${post.id} class="button-nonactive">
-                <div title="Вы можете удалить пост, созданный Вами"</div>
-                Удалить</button>`
+                  : ""
                 : ""
             }
             </div>
@@ -102,7 +102,7 @@ export function renderPostsPageComponent({ appEl }) {
         page === USER_POSTS_PAGE
           ? `<div class="posts-user-header" data-user-id=${idUser}>
           <img src=${imageUser} class="posts-user-header__user-image">
-          <p class="posts-user-header__user-name">${nameUser}</p>
+          <p class="posts-user-header__user-name">${sanitize(nameUser)}</p>
         </div>`
           : ""
       }
@@ -157,3 +157,34 @@ export function renderPostsPageComponent({ appEl }) {
     });
   }
 }
+
+const btnUp = {
+  el: document.querySelector('.btn-up'),
+  show() {
+    // удалим у кнопки класс btn-up_hide
+    this.el.classList.remove('btn-up_hide');
+  },
+  hide() {
+    // добавим к кнопке класс btn-up_hide
+    this.el.classList.add('btn-up_hide');
+  },
+  addEventListener() {
+    // при прокрутке содержимого страницы
+    window.addEventListener('scroll', () => {
+      // определяем величину прокрутки
+      const scrollY = window.scrollY || document.documentElement.scrollTop;
+      // если страница прокручена больше чем на 400px, то делаем кнопку видимой, иначе скрываем
+      scrollY > 400 ? this.show() : this.hide();
+    });
+    // при нажатии на кнопку .btn-up
+    document.querySelector('.btn-up').onclick = () => {
+      // переместим в начало страницы
+      window.scrollTo({
+        top: 0,
+        left: 0,
+        behavior: 'smooth'
+      });
+    }
+  }
+}
+btnUp.addEventListener();
