+++ index.html
@@ -10,6 +10,7 @@
   </head>
 
   <body>
+    <div class="btn-up btn-up_hide"></div>
     <div id="app"></div>
   </body>
