+++ index.js
@@ -2,8 +2,6 @@ import {
   getPosts,
   getUserPosts,
   postPosts,
-  toggleLike,
-  deletePost,
 } from "./api.js";
 import { renderAddPostPageComponent } from "./components/add-post-page-component.js";
 import { renderAuthPageComponent } from "./components/auth-page-component.js";
@@ -27,7 +25,7 @@ export let page = null;
 export let id = null;
 export let posts = [];
 
-const getToken = () => {
+export const getToken = () => {
   const token = user ? `Bearer ${user.token}` : undefined;
   return token;
 };
@@ -105,7 +103,7 @@ export const goToPage = (newPage, data) => {
   throw new Error("страницы не существует");
 };
 
-const renderApp = () => {
+export const renderApp = () => {
   const appEl = document.getElementById("app");
   if (page === LOADING_PAGE) {
     return renderLoadingPageComponent({
@@ -169,53 +167,54 @@ const renderApp = () => {
 
 goToPage(POSTS_PAGE);
 
-export const likePost = (likeId, doLike, page) => {
-  const appEl = document.getElementById("app");
-  console.log(likeId, doLike);
-  toggleLike({ token: getToken() }, likeId, doLike)
-    .then(() => {
-      getPage(page)
-        .then((newPosts) => {
-          page = POSTS_PAGE;
-          posts = newPosts;
-          renderApp();
-        })
-        .catch((error) => {
-          console.error(error);
-          goToPage(POSTS_PAGE);
-        });
-    })
-    .catch((error) => {
-      alert(error.message);
-      goToPage(AUTH_PAGE);
-    });
-};
+// export const likePost = (likeId, doLike, page) => {
+//   const appEl = document.getElementById("app");
+//   console.log(likeId, doLike);
+//   toggleLike({ token: getToken() }, likeId, doLike)
+//     .then(() => {
+//       getPage(page)
+//         .then((newPosts) => {
+//           page = POSTS_PAGE;
+//           posts = newPosts;
+//           renderApp();
+//         })
+//         .catch((error) => {
+//           console.error(error);
+//           goToPage(POSTS_PAGE);
+//         });
+//     })
+//     .catch((error) => {
+//       alert(error.message);
+//       goToPage(AUTH_PAGE);
+//     });
+// };
 
-const getPage = (page) => {
-  if (page === POSTS_PAGE) {
-    return getPosts({ token: getToken() });
-  } else {
-    return getUserPosts(id, { token: getToken() });
-  }
-};
+// const getPage = (page) => {
+//   if (page === POSTS_PAGE) {
+//     return getPosts({ token: getToken() });
+//   } else {
+//     return getUserPosts(id, { token: getToken() });
+//   }
+// };
 
-export const delPost = (delId, page) => {
-  const appEl = document.getElementById("app");
-  console.log(delId);
-  deletePost({ token: getToken() }, delId)
-  .then(() => {
-    getPage(page)
-      .then((newPosts) => {
-        page = POSTS_PAGE;
-        posts = newPosts;
-        renderApp();
-      })
-      .catch((error) => {
-        console.error(error);
-        if (error.message === "Сервер недоступен") {
-          alert("Сервер недоступен, попробуйте позже");
-        }
-        goToPage(POSTS_PAGE);
-      });
-  });
-};
+// export const delPost = (delId, page) => {
+//   const appEl = document.getElementById("app");
+//   console.log(delId);
+//   deletePost({ token: getToken() }, delId)
+//   .then(() => {
+//     getPage(page)
+//       .then((newPosts) => {
+//         page = POSTS_PAGE;
+//         posts = newPosts;
+//         renderApp();
+//       })
+//       .catch((error) => {
+//         console.error(error);
+//         if (error.message === "Сервер недоступен") {
+//           alert("Сервер недоступен, попробуйте позже");
+//         }
+//         goToPage(POSTS_PAGE);
+//       });
+//   });
+// };
+
